requirements.txt
beautifulsoup4==4.13.3
requests==2.32.3
fpdf2==2.8.2
python-docx==1.1.2
selenium==4.29.0
undetected-chromedriver==3.5.5
pathlib==1.0.1